PostView2: updated PostView2 project.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@12469 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostView2/3PointAngleTool.cpp
@@ -124,7 +124,7 @@ void C3PointAngleTool::update(bool breset)
 		{
 			FEMeshBase& mesh = *doc->GetActiveMesh();
 			const vector<FENode*> selectedNodes = doc->GetGLModel()->GetNodeSelection();
-			int N = selectedNodes.size();
+			int N = (int) selectedNodes.size();
 			int nsel = 0;
 			for (int i = 0; i<N; ++i)
 			{
