+++ PostView2/FileViewer.cpp
@@ -82,7 +82,7 @@ CFileViewer::CFileViewer(CMainWindow* pwnd, QWidget* parent) : QWidget(parent), 
 	m_filters.push_back(pair<QString, QString>("U3D files (*.u3d)"     , "*.u3d" ));
 
 	// add filters to drop down
-	int nflts = m_filters.size();
+	int nflts = (int) m_filters.size();
 	for (int i=0; i<nflts; ++i)
 	{
 		pair<QString, QString>& flt = m_filters[i];
